fix: 일부 api 요청 파일 분리
http request 요청 관련 함수를 request.ts로 분리
관심사 분리가 필요해보임
